feat(zql): push information about used columns down into `filter`
We'll want to hoist filters when getting history.
